Mayor change! Everything got optimized to avoid using N parameters on any methods but to use only a PMContext, an extension of org.jpos.transaction.Context class that comes with some helpers. With this change I have a lot more flexibility on parameter passage and it looks more like "Transaction Manager" jPOS module.
Also there is a new concept of DataAccess asociated with
entity that allows the programmer to build any data
source for his entity. Befora that, was able to do that
with contexts but now it is much cleaner, modular and
simpler. By default its defined as DataAccessDB.

There was a lots of improvements in struts actions and
error handling.